Pivot to implement a VFS driver first before tackling more complicated ones.
Also update the README.md with a description of how we plan for this to work.

Notable is the need for some type of distribution system to advertise layer
availability. For now we will use a DKV, in time we may want to use a gossip
protocol to keep with the low overhead concept.